SJRK-405: Group request handlers into separate files
