+++ index.js
@@ -11,13 +11,14 @@ https://raw.githubusercontent.com/fluid-project/sjrk-story-telling/main/LICENSE.
 
 var fluid = require("infusion");
 var kettle = require("kettle");
-require("./src/server/staticHandlerBase");
 require("./src/server/middleware/basicAuth");
 require("./src/server/middleware/saveStoryFile");
 require("./src/server/middleware/staticMiddlewareSubdirectoryFilter");
 require("./src/server/dataSource");
 require("./src/server/serverSetup");
-require("./src/server/requestHandlers");
+require("./src/server/authorsRequestHandlers");
+require("./src/server/staticRequestHandlers");
+require("./src/server/storyRequestHandlers");
 require("./src/server/validators");
 
 fluid.registerNamespace("sjrk");
