--- src/server/staticHandlerBase.js
@@ -1,22 +1,0 @@
-/*
-For copyright information, see the AUTHORS.md file in the docs directory of this distribution and at
-https://github.com/fluid-project/sjrk-story-telling/blob/main/docs/AUTHORS.md
-
-Licensed under the New BSD license. You may not use this file except in compliance with this licence.
-You may obtain a copy of the BSD License at
-https://raw.githubusercontent.com/fluid-project/sjrk-story-telling/main/LICENSE.txt
-*/
-
-"use strict";
-
-var fluid = require("infusion");
-
-// A basic static HTTP Kettle request handler that will return a 404
-fluid.defaults("sjrk.storyTelling.server.staticHandlerBase", {
-    gradeNames: "kettle.request.http",
-    invokers: {
-        handleRequest: {
-            funcName: "kettle.request.notFoundHandler"
-        }
-    }
-});
