+++ src/server/storyRequestHandlers.js
@@ -578,223 +578,3 @@ sjrk.storyTelling.server.deleteSingleFileRecoverable = function (fileToDelete, d
         fluid.log("Deleted file:", deletionPath);
     }
 };
-
-// Kettle request handler for getting Client Configuration data
-fluid.defaults("sjrk.storyTelling.server.clientConfigHandler", {
-    gradeNames: "kettle.request.http",
-    invokers: {
-        handleRequest: {
-            funcName: "sjrk.storyTelling.server.getClientConfig",
-            args: ["{arguments}.0", "{server}.options.globalConfig", "{server}.options.secureConfig"]
-        }
-    }
-});
-
-/**
- * Returns a collection of values which are "safe" to share with the client side of the application
- *
- * @param {Object} request - a Kettle request for clientConfig
- * @param {Object} globalConfig - the "global config" entry list
- * @param {Object} secureConfig - the "secure config" entry list
- */
-sjrk.storyTelling.server.getClientConfig = function (request, globalConfig, secureConfig) {
-    request.events.onSuccess.fire({
-        theme: globalConfig.theme || secureConfig.baseThemeName,
-        baseTheme: secureConfig.baseThemeName,
-        authoringEnabled: globalConfig.authoringEnabled
-    });
-};
-
-// Kettle request handler for the ui directory
-fluid.defaults("sjrk.storyTelling.server.uiHandler", {
-    gradeNames: ["sjrk.storyTelling.server.staticHandlerBase"],
-    requestMiddleware: {
-        "static": {
-            middleware: "{server}.ui"
-        }
-    }
-});
-
-// Kettle request handler for the themes directory. It looks first in the custom
-// theme directory before falling back to the base theme directory
-fluid.defaults("sjrk.storyTelling.server.themeHandler", {
-    gradeNames: ["sjrk.storyTelling.server.staticHandlerBase"],
-    requestMiddleware: {
-        // includes things like robots.txt and (in the future) favicon
-        "staticFiles": {
-            middleware: "{server}.static"
-        },
-        "baseTheme": {
-            middleware: "{server}.baseTheme",
-            priority: "before:staticFiles"
-        },
-        "currentTheme": {
-            middleware: "{server}.currentTheme",
-            priority: "before:baseTheme"
-        }
-    }
-});
-
-// Kettle request handler for the uploads directory
-fluid.defaults("sjrk.storyTelling.server.uploadsHandler", {
-    gradeNames: ["sjrk.storyTelling.server.staticHandlerBase"],
-    requestMiddleware: {
-        "static": {
-            middleware: "{server}.uploads"
-        }
-    }
-});
-
-// Kettle request handler for the node_modules directory
-fluid.defaults("sjrk.storyTelling.server.nodeModulesHandler", {
-    gradeNames: ["sjrk.storyTelling.server.staticHandlerBase"],
-    requestMiddleware: {
-        "staticFilter": {
-            middleware: "{server}.nodeModulesFilter"
-        },
-        "static": {
-            middleware: "{server}.nodeModules",
-            priority: "after:staticFilter"
-        }
-    }
-});
-
-/********************************************
- * Request Handlers for signup/login/logout *
- *******************************************/
-
-fluid.defaults("sjrk.storyTelling.server.signupHandler", {
-    gradeNames: ["kettle.request.http", "kettle.request.sessionAware"],
-    requestMiddleware: {
-        validate: {
-            middleware: "{sjrk.storyTelling.server.signupValidator}"
-        }
-    },
-    invokers: {
-        handleRequest: {
-            funcName: "sjrk.storyTelling.server.handleSignupRequest",
-            args: ["{request}", "{expressUserUtils}"]
-        }
-    }
-});
-
-/**
- * Creates a new user account and stores it to the database
- *
- * @param {Object} request - a Kettle request
- * @param {Component} expressUserUtils - an instance of `fluid.express.user.utils`
- */
-sjrk.storyTelling.server.handleSignupRequest = function (request, expressUserUtils) {
-    if (request.req.session.authorID) {
-        request.events.onError.fire({
-            statusCode: 409,
-            message: "An account is already logged in."
-        });
-        return;
-    }
-
-    if (request.req.body.password !== request.req.body.confirm) {
-        request.events.onError.fire({
-            statusCode: 400,
-            message: "The password and confirmation password do not match."
-        });
-        return;
-    }
-
-    expressUserUtils.byUsernameOrEmailReader.get({email: request.req.body.email}).then(function (body) {
-        if (body.email) {
-            request.events.onError.fire({
-                statusCode: 409,
-                message: "Account already exists for " + request.req.body.email
-            });
-            return;
-        } else {
-            var promise = expressUserUtils.createNewUser({
-                email: request.req.body.email,
-                password: request.req.body.password,
-                authorID: uuidv4()
-            });
-
-            promise.then(function (record) {
-                request.req.session.authorID = record.authorID;
-                request.events.onSuccess.fire();
-            }, function () {
-                request.events.onError.fire({
-                    statusCode: 409,
-                    message: "Unable to create account."
-                });
-            });
-        }
-    }, function () {
-        request.events.onError.fire({
-            statusCode: 500
-        });
-        return;
-    });
-};
-
-fluid.defaults("sjrk.storyTelling.server.loginHandler", {
-    gradeNames: ["kettle.request.http", "kettle.request.sessionAware"],
-    requestMiddleware: {
-        validate: {
-            middleware: "{sjrk.storyTelling.server.loginValidator}"
-        }
-    },
-    invokers: {
-        handleRequest: {
-            funcName: "sjrk.storyTelling.server.handleLoginRequest",
-            args: ["{request}", "{expressUserUtils}"]
-        }
-    }
-});
-
-/**
- * Logs in an existing user with the password and username provided in the request.
- *
- * @param {Object} request - a Kettle request
- * @param {Component} expressUserUtils - an instance of `fluid.express.user.utils`
- */
-sjrk.storyTelling.server.handleLoginRequest = function (request, expressUserUtils) {
-    if (request.req.session.authorID) {
-        request.events.onError.fire({
-            statusCode: 409,
-            message: "An account is already logged in."
-        });
-        return;
-    }
-
-    var promise = expressUserUtils.unlockUser(request.req.body.email, request.req.body.password);
-
-    promise.then(function (record) {
-        request.req.session.authorID = record.authorID;
-        request.events.onSuccess.fire();
-    }, function () {
-        request.events.onError.fire({
-            statusCode: 401
-        });
-    });
-};
-
-fluid.defaults("sjrk.storyTelling.server.logoutHandler", {
-    gradeNames: ["kettle.request.http", "kettle.request.sessionAware"],
-    invokers: {
-        handleRequest: {
-            funcName: "sjrk.storyTelling.server.handleLogoutRequest",
-            args: ["{request}"]
-        }
-    }
-});
-
-/**
- * Logs out of the session
- *
- * @param {Object} request - a Kettle request
- */
-sjrk.storyTelling.server.handleLogoutRequest = function (request) {
-    request.req.session.destroy();
-    request.events.onSuccess.fire("logout successful");
-};
-
-/****************************************
- * End Handlers for signup/login/logout *
- ***************************************/
