+++ tests/server/testServerWithMissingTheme.js
@@ -13,13 +13,14 @@ var fluid = require("infusion"),
     jqUnit = fluid.registerNamespace("jqUnit"),
     kettle = require("kettle");
 
-require("../../src/server/staticHandlerBase");
 require("../../src/server/middleware/basicAuth");
 require("../../src/server/middleware/saveStoryFile");
 require("../../src/server/middleware/staticMiddlewareSubdirectoryFilter");
 require("../../src/server/dataSource");
 require("../../src/server/serverSetup");
-require("../../src/server/requestHandlers");
+require("../../src/server/authorsRequestHandlers");
+require("../../src/server/staticRequestHandlers");
+require("../../src/server/storyRequestHandlers");
 require("../../src/server/validators");
 require("./utils/serverTestUtils.js");
